Rename component and partial fix to MapLibre tile error
