+++ frontend/app/components/entity/EntityIconMobile.vue
